Crate torch_sys_cuda

Re-exports

pub use nccl_stubs as nccl;

Modules

cuda

A companion to the torch-sys crate that provides bindings for CUDA-specific functionality from libtorch. This crate is separated out to make it easier for clients who want to avoid compiling CUDA code.

nccl_common

Types shared between the real NCCL bindings (nccl) and the CPU-only stubs (nccl_stubs).

nccl_stubs

Stub types for NCCL when the cuda feature is disabled.